Conflicts:
	cave/com.raytheon.viz.feature.awips.developer/feature.xml

Former-commit-id: 1077e3c2735fdeab1766fc93d7e53c6a4eb98261 [formerly 378db03f9724b562a39ab4f2847b6cb2b5901eab] [formerly 716c1c8daeeb143fdb46a82f2c64c8ebd288b28b] [formerly 716c1c8daeeb143fdb46a82f2c64c8ebd288b28b [formerly 4535f432ceb0de1b4229a4a70169ec059b268345]] [formerly 1077e3c2735fdeab1766fc93d7e53c6a4eb98261 [formerly 378db03f9724b562a39ab4f2847b6cb2b5901eab] [formerly 716c1c8daeeb143fdb46a82f2c64c8ebd288b28b] [formerly 716c1c8daeeb143fdb46a82f2c64c8ebd288b28b [formerly 4535f432ceb0de1b4229a4a70169ec059b268345]] [formerly dafec4c8fe48dfb5dbc1c1530cc9d721e6a52689 [formerly 716c1c8daeeb143fdb46a82f2c64c8ebd288b28b [formerly 4535f432ceb0de1b4229a4a70169ec059b268345] [formerly dafec4c8fe48dfb5dbc1c1530cc9d721e6a52689 [formerly f2c656a4fe6d4b82f691e10d8dc49dcbbcff5adc]]]]]
Former-commit-id: dafec4c8fe48dfb5dbc1c1530cc9d721e6a52689
Former-commit-id: ee0bdd133ea881e84f959892ceb335134b038829 [formerly fc36bb9df08d826f916f6f463673a5b88edd7d42] [formerly 46d475cc451d2590ba6aa4afe0a98971485df135] [formerly 871e61d152ae051fb99e63e9eedaf9822f632ffe [formerly c18a3f184c31a1cd39603e8d4749a241d056a0f4] [formerly 46d475cc451d2590ba6aa4afe0a98971485df135 [formerly 0a6f1170fe8e661fde04e8150f52c272e0743298]]]
Former-commit-id: 74d39546ae650783dfba328fe6bd9d7629e78d83 [formerly 535947695d9fa08de45fb144a87af03dfe031c9d] [formerly 27d75ed5cf32736d4f04e32809bf6ec88ecd5153 [formerly d42d3aa9915973fd9e9b032982a275b125d5f948]]
Former-commit-id: 27d75ed5cf32736d4f04e32809bf6ec88ecd5153
Former-commit-id: eeed7718600f64574c74e77a3f80d052c82ce522

+++ deltaScripts/14.3.1/DR16693/setupDict.sh
@@ -15,50 +15,61 @@
 #clean up platform dictionary
 #
 
+cmddir=$(dirname $_)
+FNAME="spelldict"
 user=$(whoami)
 host=$(hostname)
+stddict=$cmddir/$FNAME.std
+
+[ ! -f $stddict ] && {
+    echo Error: the standard dictionary \"$stddict\" not found.
+    exit 1
+}
 
 edex_user_dir=/awips2/edex/data/utility/cave_static/user/
 cave_etc=/awips2/cave/etc
 run_type=0
-FNAME="spelldict"
 
 clean () {
     lines=`cat $1 |wc -l`
     size=`cat $1 |wc -c`
     MSG="$1, size=$size, #line=$lines:"
     LASTWD=$(grep 'zymurgy' $1)
-    if [ $size -eq 1290760 ]
+
+    if [ $run_type == 1 ] 
     then
-        remove $1
-#    elif [ $lines -gt 135553 ]  
-#    then
-#        [ $run_type == 1 ] && (cp $1 "$1.bak"; 
-#            sed -n "135554,${lines}p" "$1.bak" > $1)
-#        let "newlines=${lines}-135553"
-#        echo $MSG modified, \#line=$(( lines-135553 ))
-    elif [ "$LASTWD" ]
-    then
-        line=$(sed -n "/^$LASTWD/=" $1)
-#        echo line=$line
-        [ $run_type == 1 ] && (cp -p $1 "$1.bak"; sed "1, /^$LASTWD/d" "$1.bak" > $1)
-        echo $MSG "modified, #line=$(( lines-line ))"
+        cp -p $1 "$1.bak"
+        cp $stddict $1
+        if [ "$LASTWD" ]
+        then
+            sed "1, /^$LASTWD/d" "$1.bak" >> $1
+        else
+            cat $1.bak >> $1
+        fi
+        echo "$MSG modified, #line=$(cat $1 |wc -l)" 
     else
-        echo $MSG unchanged
+        echo $MSG "modified"
     fi
 }
 
 remove () {
-    lines=`cat $1 |wc -l`
-    size=`cat $1 |wc -c`
-    if [ $run_type == 1 ]
+    if [ $run_type == 1 ] && [ -f $1 ]
     then
-        cp -p $1 "$1.bak"
+        mv  $1 "$1.bak"
         [[ $1 == ${cave_etc}* ]] && cat /dev/null > $1 || rm -f $1
     fi
         
-    action=$([[ $1 == ${cave_etc}* ]] && echo emptied || echo removed )
-    echo "$1, size=$size, #line=$lines: $action"
+    echo "$1, removed"
+}
+
+create () {
+    [ $run_type == 1 ] && ( 
+        cp $stddict $1
+        chown awips $1
+        chgrp fxalpha $1
+        chmod 644 $1
+    )
+    echo "$1, created the standard dictionary"
 }
 
 usage () {
@@ -110,11 +121,7 @@ then
     then
         clean $f 
     else
-        cat /dev/null > $f
-        chown awips $f
-        chgrp fxalpha $f
-        chmod 644 $f
-        echo $f: created, size=0
+        create $f
     fi
 fi
